backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
flask-jwt-extended==4.7.0
marshmallow==3.23.2
hypothesis==6.122.3
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
scikit-learn==1.8.0
setuptools>=65.0.0
pyarrow>=18.0.0
werkzeug==3.1.3
pytest==8.3.4
requests==2.32.3
twilio==9.0.4
python-dotenv==1.0.0
